requirements.txt
asgiref==3.7.2
attrs==23.1.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
config==0.5.1
Django==4.2.4
django-background-tasks-updated==1.2.7
exceptiongroup==1.1.3
fake-useragent==1.2.1
flake8==6.1.0
flake8-bugbear==23.7.10
flake8-builtins==2.1.0
flake8-import-order==0.18.2
h11==0.14.0
idna==3.4
mccabe==0.7.0
outcome==1.2.0
pep8-naming==0.13.3
pycodestyle==2.11.0
pycparser==2.21
pyflakes==3.1.0
PySocks==1.7.1
pyTelegramBotAPI==4.14.0
python-dateutil==2.8.2
python-decouple==3.8
requests==2.31.0
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
telebot==0.0.5
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
urllib3==2.0.4
wsproto==1.2.0
